+++ tests/testthat/test-summarise.r
@@ -758,7 +758,7 @@ test_that("summarise fails gracefully on raw columns (#1803)", {
   df <- data_frame(a = 1:3, b = as.raw(1:3))
   expect_error(
     summarise(df, c = b[[1]]),
-    'Unsupported type RAWSXP for column "c"'
+    "Column `c` must be a vector, not a raw vector"
   )
 })